change dataset.py

+++ dataset.py
@@ -260,7 +260,7 @@ class ToTensor(object):
         if 'obj_cam' in sample:
             obj_cam = sample['obj_cam']
             sample['obj_cam'] = \
-                transforms.functional.to_tensor(obj_cam).squeeze().float()
+                transforms.functional.to_tensor(obj_cam).squeeze().long()
 
             # in label bg class is not included
             sample['obj_label'] = torch.cat([torch.tensor([0.]), sample['obj_label']])
